v0.28d

+++ setup.sh
@@ -40,12 +40,12 @@ function clean() {
 
 function re() {
 	RE=1
-	echo "${green}Deleting dangling (untagged) docker images.${reset}"
+	echo -e "${red}Deleting ${cyan}dangling ${green}(untagged) docker images.${reset}"
 	docker image rm $(docker images -q --filter "dangling=true")
 }
 
 function k_del() {
-	echo "${green}Deleting minikube pods, deployments and services.${reset}"
+	echo -e "${red}Deleting ${cyan}minikube pods${green}, ${cyan}deployments${green} and ${cyan}services${green}.${reset}"
 	kubectl delete pod --all
 	kubectl delete deployment --all
 	kubectl delete service --all
@@ -54,7 +54,7 @@ function k_del() {
 }
 
 function k_apply() {
-	echo "${green}Applying all yml files${reset}"
+	echo -e "${green_b}Applying${green} all ${cyan}yml ${green}files${reset}"
 	kubectl apply -f srcs/grafana.yml
 	kubectl apply -f srcs/influxdb.yml
 	sed -i "" "s/__IP_VAR__/$IP_VAR/g" ./srcs/nginx.yml
@@ -64,43 +64,46 @@ function k_apply() {
 }
 
 function d_del() {
-	echo "${green}Deleting docker images: nginx-services, grafana and influxdb${reset}"
+	echo -e "${red}Deleting${cyan} docker images${green}: nginx-services, grafana and influxdb${reset}"
 	docker image rm -f nginx-services
 	docker image rm -f grafana
 	docker image rm -f influxdb
 }
 
 function d_nginx() {
-	if [ $(containsElement "--d_del" $@) -eq 0 ]
+	if [ $(containsElement "--d_del" $@) -eq 0 ] && [ $RE -eq 1 ]
 	then
-		echo "${green}Deleting nginx docker image${reset}"
+		echo -e "${red}Deleting${cyan} nginx docker image${reset}"
 		docker image rm -f nginx-services
 	fi
 	sed -i "" "s/__IP_VAR__/$IP_VAR/g" ./srcs/nginx/srcs/index.html
-	echo "${green}BUILDING NGINX${reset}"
+	echo -e "${green_b}BUILDING${cyan} nginx${reset}"
 	docker build -t nginx-services srcs/nginx/.
+		echo -e "${red}Docker return val pre-wait: $? ${reset}"
 	wait
+		echo -e "${red}Docker return val: $? ${reset}"
 	sed -i "" "s/$IP_VAR/__IP_VAR__/g" ./srcs/nginx/srcs/index.html
 }
 
 function d_grafana() {
 	if [ $(containsElement "--d_del" $@) -eq 0 ] && [ $RE -eq 1 ]
 	then
-		echo "${green}Deleting grafana docker image${reset}"
+		echo -e "${red}Deleting${cyan} grafana docker image${reset}"
 		docker image rm -f grafana
 	fi
-	echo "${green}BUILDING GRAFANA${reset}"
+	echo "${green_b}BUILDING${cyan} grafana${reset}"
 	docker build -t grafana srcs/grafana/.
+		echo -e "${red}Docker return val: $? ${reset}"
 	wait
 }
 
 function d_influxdb() {
-	if [ $(containsElement "--d_del" $@) -eq 0 ]
+	if [ $(containsElement "--d_del" $@) -eq 0 ] && [ $RE -eq 1 ]
 	then
-		echo "${green}Deleting influxdb docker image${reset}"
+		echo "${red}Deleting${cyan} influxdb docker image${reset}"
 		docker image rm -f influxdb
 	fi
-	echo "${green}BUILDING INFLUXDB${reset}"
+	echo "${green_b}BUILDING${cyan} influxdb${reset}"
 	docker build -t influxdb srcs/influxdb/.
 	wait
 }
@@ -117,11 +120,53 @@ function containsElement() {
 	fi
 }
 
+function error() {
+	echo -e "${green}An ${red}error${green} was encountered when building the docker image for: ${cyan}$1${green}."
+	docker image ls > /dev/null
+	local ret1=$?
+	local ret2
+	if [ $(docker image ls | grep "k8s.gcr.io" -ic) -gt 0 ]
+	then
+		ret2=1
+	else
+		ret2=0
+	fi
+	echo -e "${red}ret1 is: $ret1${reset}"
+	echo -e "${red}ret2 is: $ret2${reset}"
+	
+	if [ $ret1 -eq 1 ]
+	then
+		echo -e "${green}A simple check reveals that ${cyan}Docker${green} is not running, or that the Docker daemon is otherwise not responding.${reset}"
+		echo -e "Check whether this is the case, initiate Docker and re-run this setup."
+ # CONTINUE HERE ************************
+	while [ "$input" != "n" ] && [ "$input" != "y" ] && [ "$input" != "N" ] && [ "$input" != "Y" ]
+		do
+			if [ $ret -eq 1 ]
+			then
+				read -n1 -p "${green}A simple check reveals that ${cyan}Docker${green} is not running, or that the Docker daemon is otherwise not responding.
+			read -n1 -p "${green}An ${red}error${green} was encountered when building the docker image for: ${cyan}$1${green}."
+		   
+			Is Do you want to setup${cyan} $1${green}?${green_d}$3 ${magenta}(y/n)${blue} " input
+			echo "${reset}"
+		done
+	if [ "$input" = "y" ] || [ "$input" = "Y" ]
+		then
+		echo "${green}Executing: ${cyan}$2${reset}"
+		bash $2
+		echo "${green_b}Done${reset}"
+	fi
+	if [ "$input" = "n" ] || [ "$input" = "N" ]
+		then
+		echo "Skipping ${cyan}$1${reset}"
+	fi
+	input=""
+}
+
 function setup() {
 	local e
 	if [ $# -eq 0 ]
 	then
-		echo "No args, default behaviour"
+		echo "${green}No args, ${blue}default${green} behaviour${reset}"
 		k_del
 		d_nginx
 		d_influxdb
@@ -130,13 +175,10 @@ function setup() {
 	else
 		for e in "${KEYW[@]}"
 		do
-			#$echo "checking keyword: $e"
 			if [ $(containsElement $e $@) -eq 1 ]
 			then
-				#echo "found: $e"
 				$(echo $e | cut -c 3-) # remove the "--" from the keyword and execute function with corresponding name
 			#else
-				#echo "nothing so far..."
 			fi
 		done
 	fi
